split out backlink reply mapping into its own file (so it can be used on its own)
